+++ pydmd/__init__.py
@@ -36,4 +36,4 @@ from .rdmd import RDMD
 from .havok import HAVOK
 from .bopdmd import BOPDMD
 from .pidmd import PiDMD
-from .pre_post_processing import PrePostProcessingDMD
+from .preprocessing import PrePostProcessingDMD
