+++ pydmd/preprocessing/__init__.py
@@ -3,3 +3,4 @@ PyDMD preprocessing init.
 """
 
 from .pre_post_processing import PrePostProcessingDMD
+from .zero_mean import zero_mean_preprocessing
