+++ pydmd/preprocessing/zero_mean.py
@@ -1,0 +1,23 @@
+from typing import Dict
+import numpy as np
+from pydmd.dmdbase import DMDBase
+from pydmd.preprocessing import PrePostProcessingDMD
+
+
+def zero_mean_preprocessing(dmd: DMDBase):
+    """
+    Zero-mean pre-processing.
+
+    :param dmd: DMD instance to be wrapped.
+    :type dmd: DMDBase
+    """
+    return PrePostProcessingDMD(dmd, _pre, _post)
+
+
+def _pre(state: Dict, X: np.ndarray, **kwargs):
+    state["mean"] = np.mean(X)
+    return X - state["mean"], *kwargs.values()
+
+
+def _post(state: Dict, X: np.ndarray):
+    return X + state["mean"]
