+++ tests/preprocessing/test_pre_post_processing.py
@@ -1,9 +1,5 @@
-import numpy as np
 from unittest.mock import patch
-from pydmd.pre_post_processing import (
-    PrePostProcessingDMD,
-    zero_mean_preprocessing,
-)
+from pydmd.preprocessing import PrePostProcessingDMD
 
 _kwargs = {"param1": "value1", "param2": "value2"}
 
@@ -23,7 +19,9 @@ def test_pre_processing(mocker):
 
     X = mocker.Mock()
     mock_dict = mocker.Mock()
-    with patch("pydmd.pre_post_processing.dict", return_value=mock_dict):
+    with patch(
+        "pydmd.preprocessing.pre_post_processing.dict", return_value=mock_dict
+    ):
         pdmd.fit(X, **_kwargs)
 
     pre.assert_called_once_with(mock_dict, X, **_kwargs)
@@ -62,29 +60,3 @@ def test_pre_processing_default_postprocessing(mocker):
     pdmd.fit(X)
 
     assert pdmd.reconstructed_data is partial_output
-
-
-def test_zero_mean(mocker):
-    dmd = mocker.Mock()
-    pdmd = zero_mean_preprocessing(dmd)
-
-    X = np.array([1, 2, 3], dtype=float)
-    pdmd.fit(X)
-    fit_call_args = dmd.fit.call_args_list
-    assert len(fit_call_args) == 1
-    assert (fit_call_args[0][0][0] == [-1, 0, 1]).all()
-
-    dmd.reconstructed_data = np.zeros(3, dtype=float)
-    assert (pdmd.reconstructed_data == [2, 2, 2]).all()
-
-
-def test_zero_mean_with_kwargs(mocker):
-    dmd = mocker.Mock()
-    pdmd = zero_mean_preprocessing(dmd)
-
-    X = np.array([1, 2, 3], dtype=float)
-    pdmd.fit(X, **_kwargs)
-    fit_call_args = dmd.fit.call_args_list
-    assert len(fit_call_args) == 1
-    assert (fit_call_args[0][0][0] == [-1, 0, 1]).all()
-    assert fit_call_args[0][0][1:] == tuple(_kwargs.values())
